tools/syz-reprolist: switch to dashboard/api.Client
